+++ codegen/codegen.sh
@@ -40,29 +40,38 @@ git clone --quiet \
     https://github.com/sigstore/rekor \
     "${rekor_dir}"
 
-# NOTE(ww): For whatever reason, the POST endpoint doesn't work. Instead,
-# we tell the converter to retrieve the OpenAPI YAML URL itself.
-curl -X 'GET' \
-  "https://converter.swagger.io/api/convert?url=https%3A%2F%2Fraw.githubusercontent.com%2Fsigstore%2Frekor%2F${rekor_ref}%2Fopenapi.yaml" \
-  -H 'accept: application/json' \
-  > "${rekor_dir}/openapi.json"
-
-datamodel-codegen \
-    --input "${rekor_dir}/openapi.json" \
-    --input-file-type openapi \
-    --target-python-version 3.8 \
-    --snake-case-field \
-    --capitalize-enum-members \
-    --field-constraints \
-    --use-schema-description \
-    --use-subclass-enum \
-    --disable-timestamp \
-    --reuse-model \
-    --use-default-kwarg \
-    --allow-population-by-field-name \
-    --strict-types str bytes int float bool \
-    --output-model-type pydantic_v2.BaseModel \
-    --output "${pkg_dir}/_internal.py"
+# NOTE(ww): This is a miserable hack: ideally we'd use Rekor's top-level
+# `openapi.yaml`, but it's written in OpenAPI 2.0 (because go-swagger only
+# supports 2.0). Meanwhile, `datamodel-codegen` only supports OpenAPI 3.0+.
+# Neither can be trivially upgraded. As a stop-gap, we poke through each
+# internal JSON Schema definition used by Rekor and generate them one-by-one.
+#
+# See:
+#  * https://github.com/go-swagger/go-swagger/issues/1122
+#  * https://github.com/koxudaxi/datamodel-code-generator/issues/1590
+#  * https://github.com/sigstore/rekor/issues/1729
+mkdir -p "${pkg_dir}/_internal"
+touch "${pkg_dir}/_internal/__init__.py"
+rekor_types=(alpine cose dsse hashedrekord helm intoto jar rekord rfc3161 rpm tuf)
+for type in "${rekor_types[@]}"; do
+    dbg "generating models for Rekor type: ${type}"
+    datamodel-codegen \
+        --input "${rekor_dir}/pkg/types/${type}/${type}_schema.json" \
+        --input-file-type jsonschema \
+        --target-python-version 3.8 \
+        --snake-case-field \
+        --capitalize-enum-members \
+        --field-constraints \
+        --use-schema-description \
+        --use-subclass-enum \
+        --disable-timestamp \
+        --reuse-model \
+        --use-default-kwarg \
+        --allow-population-by-field-name \
+        --strict-types str bytes int float bool \
+        --output-model-type pydantic_v2.BaseModel \
+        --output "${pkg_dir}/_internal/${type}.py"
+done
 
 # Cap it off by auto-reformatting.
 make -C "${here}/.." reformat
