+++ sigstore_rekor_types/_internal/dsse.py
@@ -1,0 +1,122 @@
+# generated by datamodel-codegen:
+
+from __future__ import annotations
+
+from enum import Enum
+from typing import List, Optional
+
+from pydantic import BaseModel, ConfigDict, Field, RootModel, StrictStr
+
+
+class ProposedContent(BaseModel):
+    model_config = ConfigDict(
+        populate_by_name=True,
+    )
+    envelope: StrictStr = Field(
+        ...,
+        description="DSSE envelope specified as a stringified JSON object",
+    )
+    verifiers: List[str] = Field(
+        ...,
+        description=(
+            "collection of all verification material (e.g. public keys or certificates) used to"
+            " verify signatures over envelope's payload, specified as base64-encoded strings"
+        ),
+        min_length=1,
+    )
+
+
+class Signature(BaseModel):
+    """a signature of the envelope's payload along with the verification material for the signature."""
+
+    model_config = ConfigDict(
+        populate_by_name=True,
+    )
+    signature: StrictStr = Field(..., description="base64 encoded signature of the payload")
+    verifier: str = Field(
+        ...,
+        description=(
+            "verification material that was used to verify the corresponding signature, specified"
+            " as a base64 encoded string"
+        ),
+    )
+
+
+class Algorithm(str, Enum):
+    """The hashing function used to compute the hash value."""
+
+    SHA256 = "sha256"
+
+
+class EnvelopeHash(BaseModel):
+    """Specifies the hash algorithm and value encompassing the entire envelope sent to Rekor."""
+
+    model_config = ConfigDict(
+        populate_by_name=True,
+    )
+    algorithm: Algorithm = Field(
+        ...,
+        description="The hashing function used to compute the hash value",
+    )
+    value: StrictStr = Field(
+        ...,
+        description="The value of the computed digest over the entire envelope",
+    )
+
+
+class PayloadHash(BaseModel):
+    """Specifies the hash algorithm and value covering the payload within the DSSE envelope."""
+
+    model_config = ConfigDict(
+        populate_by_name=True,
+    )
+    algorithm: Algorithm = Field(
+        ...,
+        description="The hashing function used to compute the hash value",
+    )
+    value: StrictStr = Field(
+        ...,
+        description="The value of the computed digest over the payload within the envelope",
+    )
+
+
+class DsseV001Schema(BaseModel):
+    """Schema for DSSE envelopes."""
+
+    model_config = ConfigDict(
+        populate_by_name=True,
+    )
+    proposed_content: Optional[ProposedContent] = Field(default=None, alias="proposedContent")
+    signatures: Optional[List[Signature]] = Field(
+        default=None,
+        description=(
+            "extracted collection of all signatures of the envelope's payload; elements will be"
+            " sorted by lexicographical order of the base64 encoded signature strings"
+        ),
+        min_length=1,
+    )
+    envelope_hash: Optional[EnvelopeHash] = Field(
+        default=None,
+        alias="envelopeHash",
+        description=(
+            "Specifies the hash algorithm and value encompassing the entire envelope sent to Rekor"
+        ),
+    )
+    payload_hash: Optional[PayloadHash] = Field(
+        default=None,
+        alias="payloadHash",
+        description=(
+            "Specifies the hash algorithm and value covering the payload within the DSSE envelope"
+        ),
+    )
+
+
+class DsseSchema(RootModel[DsseV001Schema]):
+    model_config = ConfigDict(
+        populate_by_name=True,
+    )
+    root: DsseV001Schema = Field(
+        ...,
+        description="log entry schema for dsse envelopes",
+        title="DSSE Schema",
+    )
