+++ sigstore_rekor_types/_internal/rfc3161.py
@@ -1,0 +1,37 @@
+# generated by datamodel-codegen:
+
+from __future__ import annotations
+
+from pydantic import BaseModel, ConfigDict, Field, RootModel
+
+
+class Tsr(BaseModel):
+    """Information about the tsr file associated with the entry."""
+
+    model_config = ConfigDict(
+        populate_by_name=True,
+    )
+    content: str = Field(
+        ...,
+        description="Specifies the tsr file content inline within the document",
+    )
+
+
+class Rfc3161V001Schema(BaseModel):
+    """Schema for RFC3161 entries."""
+
+    model_config = ConfigDict(
+        populate_by_name=True,
+    )
+    tsr: Tsr = Field(..., description="Information about the tsr file associated with the entry")
+
+
+class TimestampSchema(RootModel[Rfc3161V001Schema]):
+    model_config = ConfigDict(
+        populate_by_name=True,
+    )
+    root: Rfc3161V001Schema = Field(
+        ...,
+        description="Schema for RFC 3161 timestamp objects",
+        title="Timestamp Schema",
+    )
