+++ sigstore_rekor_types/_internal/rpm.py
@@ -1,0 +1,37 @@
+# generated by datamodel-codegen:
+
+from __future__ import annotations
+
+from pydantic import BaseModel, ConfigDict, Field, RootModel
+
+
+class PublicKey(BaseModel):
+    """The PGP public key that can verify the RPM signature."""
+
+    model_config = ConfigDict(
+        populate_by_name=True,
+    )
+    content: str = Field(
+        ...,
+        description="Specifies the content of the public key inline within the document",
+    )
+
+
+class RpmV001Schema(BaseModel):
+    """Schema for RPM entries."""
+
+    model_config = ConfigDict(
+        populate_by_name=True,
+    )
+    public_key: PublicKey = Field(
+        ...,
+        alias="publicKey",
+        description="The PGP public key that can verify the RPM signature",
+    )
+
+
+class RpmSchema(RootModel[RpmV001Schema]):
+    model_config = ConfigDict(
+        populate_by_name=True,
+    )
+    root: RpmV001Schema = Field(..., description="Schema for RPM objects", title="RPM Schema")
